+++ confignode/src/main/java/org/apache/iotdb/confignode/manager/load/balancer/RouteBalancer.java
@@ -21,9 +21,11 @@ package org.apache.iotdb.confignode.manager.load.balancer;
 import org.apache.iotdb.common.rpc.thrift.TConsensusGroupId;
 import org.apache.iotdb.common.rpc.thrift.TConsensusGroupType;
 import org.apache.iotdb.common.rpc.thrift.TRegionReplicaSet;
+import org.apache.iotdb.commons.cluster.NodeStatus;
 import org.apache.iotdb.confignode.conf.ConfigNodeDescriptor;
 import org.apache.iotdb.confignode.manager.IManager;
-import org.apache.iotdb.confignode.manager.load.LoadManager;
+import org.apache.iotdb.confignode.manager.NodeManager;
+import org.apache.iotdb.confignode.manager.PartitionManager;
 import org.apache.iotdb.confignode.manager.load.balancer.router.IRouter;
 import org.apache.iotdb.confignode.manager.load.balancer.router.LazyGreedyRouter;
 import org.apache.iotdb.confignode.manager.load.balancer.router.LeaderRouter;
@@ -84,9 +86,9 @@ public class RouteBalancer {
       case SchemaRegion:
         if (policy.equals(leaderPolicy)) {
           return new LeaderRouter(
-              getLoadManager().getAllLeadership(), getLoadManager().getAllLoadScores());
+              getPartitionManager().getAllLeadership(), getNodeManager().getAllLoadScores());
         } else {
-          return new LoadScoreGreedyRouter(getLoadManager().getAllLoadScores());
+          return new LoadScoreGreedyRouter(getNodeManager().getAllLoadScores());
         }
       case DataRegion:
       default:
@@ -95,19 +97,24 @@ public class RouteBalancer {
             .getDataRegionConsensusProtocolClass()
             .equals(ConsensusFactory.MultiLeaderConsensus)) {
           // Latent router for MultiLeader consensus protocol
-          lazyGreedyRouter.updateUnknownDataNodes(getLoadManager().getUnknownDataNodes());
+          lazyGreedyRouter.updateUnknownDataNodes(
+              getNodeManager().filterDataNodeThroughStatus(NodeStatus.Unknown));
           return lazyGreedyRouter;
         } else if (policy.equals(leaderPolicy)) {
           return new LeaderRouter(
-              getLoadManager().getAllLeadership(), getLoadManager().getAllLoadScores());
+              getPartitionManager().getAllLeadership(), getNodeManager().getAllLoadScores());
         } else {
-          return new LoadScoreGreedyRouter(getLoadManager().getAllLoadScores());
+          return new LoadScoreGreedyRouter(getNodeManager().getAllLoadScores());
         }
     }
   }
 
-  private LoadManager getLoadManager() {
-    return configManager.getLoadManager();
+  private NodeManager getNodeManager() {
+    return configManager.getNodeManager();
+  }
+
+  private PartitionManager getPartitionManager() {
+    return configManager.getPartitionManager();
   }
 
   public Map<TConsensusGroupId, TRegionReplicaSet> getRouteMap() {
